refactor: Update room component UI to display video streams
Refactor the room component HTML to update the UI for displaying video streams. The previous implementation used a grid layout, but it has been changed to use a flex layout with a wrapping container. Each video stream is now displayed in a self-stretching flex item with a minimum width of 300px and a minimum height of 200px. The background color of the video container is set to black, and the video element itself has a maximum width and height of the container. This change improves the visual presentation of the video streams in the room component.